Moved AssessmentWindow to client since it will be the return model for the endpoint
