+++ service/src/main/java/tds/config/repositories/ExamWindowQueryRepository.java
@@ -3,8 +3,8 @@ package tds.config.repositories;
 import java.util.List;
 import java.util.Optional;
 
+import tds.config.AssessmentWindow;
 import tds.config.model.AssessmentProperties;
-import tds.config.model.AssessmentWindow;
 
 public interface ExamWindowQueryRepository {
     List<AssessmentWindow> findCurrentExamWindows(String clientName, String assessmentId, int shiftWindowStart, int shiftWindowEnd, int sessionType);
