+++ service/src/main/java/tds/config/services/ConfigService.java
@@ -3,9 +3,9 @@ package tds.config.services;
 import java.util.List;
 import java.util.Optional;
 
+import tds.config.AssessmentWindow;
 import tds.config.ClientSystemFlag;
 import tds.config.ClientTestProperty;
-import tds.config.model.AssessmentWindow;
 import tds.config.model.ExamWindowProperties;
 
 /**
