+++ service/src/test/java/tds/config/services/impl/ConfigServiceImplIntegrationTests.java
@@ -9,9 +9,9 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import java.util.List;
 import java.util.Optional;
 
+import tds.config.AssessmentWindow;
 import tds.config.ClientSystemFlag;
 import tds.config.ClientTestProperty;
-import tds.config.model.AssessmentWindow;
 import tds.config.model.ExamWindowProperties;
 import tds.config.services.ConfigService;
